requirements.txt
albumentations==0.5.2
braceexpand==0.1.7
detectron2==0.6
easydict==1.9
imageio==2.31.1
imgaug==0.4.0
joblib==1.3.1
kornia==0.5.0
loguru==0.6.0
matplotlib==3.7.2
memory_profiler==0.61.0
numpy==1.25.1
omegaconf==2.1.2
opencv_python==4.8.0.74
opencv_python_headless==4.8.0.74
pandas==2.0.3
Pillow==10.0.0
pytorch_lightning==1.2.9
PyYAML==6.0.1
scikit_learn==1.3.0
scipy==1.11.1
six==1.16.0
skimage==0.0
sync_batchnorm==0.0.1
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
webdataset==0.2.48
